move KernelInterface integration code to src
