+++ Project.toml
@@ -20,6 +20,7 @@ Functors = "d9f16b24-f501-4c13-a1f2-28368ffc5196"
 GPUArraysCore = "46192b85-c4d5-4398-a991-12ede77f4527"
 GPUCompiler = "61eb1bfa-7361-4325-ad38-22787b887f55"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+KernelInterface = "4ee993da-d684-4d17-a7dd-4e58e78d92bf"
 LLVM = "929cbde3-209d-540e-8aea-75f648917ca0"
 LLVMOpenMP_jll = "1d63c593-3942-5779-bab2-d838dc0a180e"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
@@ -137,6 +138,7 @@ GPUCompiler = "1.17"
 IrrationalConstants = "0.1, 0.2"
 JSON = "1.4"
 KernelAbstractions = "0.9.30"
+KernelInterface = "0.1.0"
 LLVM = "9.4.6"
 LLVMOpenMP_jll = "18.1.7, 22.1"
 Libdl = "1.10"
